% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_sa_tc_log.r
\docType{data}
\name{cc_sa_tc_log}
\alias{cc_sa_tc_log}
\title{Seasonally Adjusted Series for Continued Unemployment Claims, weekly (automatic outlier identification, TC outlier specified, log)}
\format{
An xts time series object for seasonally adjusted CC from January 31, 2004 to January 29, 2022
}
\usage{
cc_sa_tc_log
}
\description{
A time series object with seasonal adjustment generated for weekly continued unemployment
claims from a model fit to the logged series that includes automatically identified outliers with a temporary
change outlier (TC) specified for the 13th week of 2020
}
\keyword{datasets}
