% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_week_xts.r
\docType{data}
\name{cc_week_xts}
\alias{cc_week_xts}
\title{Week vector for Continued Unemployment Claims, weekly (cc), as an xts time series object}
\format{
An xts time series object for CC from January 31, 2004 to January 29, 2022
}
\source{
\url{https://www.dol.gov/ui/data.pdf}
}
\usage{
cc_week_xts
}
\description{
A xts time series object containing the week number for every observation of the weekly continued unemployment claims released by the Department of Labor
}
\keyword{datasets}
