% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_air_components.r
\name{gen_air_components}
\alias{gen_air_components}
\title{Generate signal extraction components}
\usage{
gen_air_components(
  this_est,
  this_decomp,
  this_xtype,
  this_x = NULL,
  this_log = TRUE,
  this_stde = FALSE,
  tc2trend = FALSE,
  remove_user = FALSE
)
}
\arguments{
\item{this_est}{A list object generated by the estimation procedure of the fractional airline model}

\item{this_decomp}{A list object generated by the decomposition procedure of the fractional airline model}

\item{this_xtype}{Character vector; Type of regressor used for including regression effects into components (\code{'hol'} for holiday, \code{'ao'} for point outliers, \code{'ls'} for level change outliers).}

\item{this_x}{Character vector; regression. Default: use the regression matrix in \code{this_est$model$X}.}

\item{this_log}{Logical scalar; set to TRUE if the log transformation is used in the decomposition, FALSE otherwise. Default: TRUE}

\item{this_stde}{Logical scalar; component standard error included in \code{this_decomp}. Default: FALSE.}

\item{tc2trend}{Logical scalar; TC outliers included in trend component if TRUE. Default: FALSE, TC outliers included in irregular.}

\item{remove_user}{Logical scalar; remove user-defined regression factors from seasonally adjusted series if TRUE. Default: FALSE, user defined regression factors included in seasonally adjusted series.}
}
\value{
List of components from the fractional airline decomposition, partioning regression effects to different components as in \code{X-13ARIMA-SEATS}
}
\description{
Generates components from the decomposition of the fractional airline model R routines, partioning regression effects to different components as in \code{X-13ARIMA-SEATS}
}
\examples{
ic_est <-
   rjd3highfreq::fractionalAirlineEstimation(ic_obs, periods=c(365.25/7),
                                      x=ic_default_matrix)
ic_decomp <-
   rjd3highfreq::fractionalAirlineDecomposition(ic_est$model$linearized,
                                         365.25/7, stde = TRUE)
ic_xtype <- c(rep('hol', 13), rep('ao', 40))
ic_comp <- 
   gen_air_components(ic_est, ic_decomp,
                      this_xtype = ic_xtype,
                      this_log = FALSE, 
                      this_stde = TRUE)
ic_comp_tis <- lapply(ic_comp, function(x)
   try(tis::tis(x, start = ic_start, tif = 'wsaturday')))
}
