% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_air_projected_factors.r
\name{gen_air_projected_factors}
\alias{gen_air_projected_factors}
\title{Generate projected seasonal (and holiday) adjustment factors}
\usage{
gen_air_projected_factors(
  this_est,
  this_decomp,
  this_xtype,
  this_x = NULL,
  this_log = TRUE,
  nfcasts = 104,
  return_series = "combined",
  return_xts = FALSE
)
}
\arguments{
\item{this_est}{A list object generated by the estimation procedure of the fractional airline model; the regression matrix should not be extended with forecasts.}

\item{this_decomp}{A list object generated by the decomposition procedure of the fractional airline model; the \code{nfcasts} option should have been specified.}

\item{this_xtype}{Character vector; Type of regressor used for including regression effects into components (\code{'hol'} for holiday, \code{'ao'} for point outliers, \code{'ls'} for level change outliers).}

\item{this_x}{Character vector; regression matrix.}

\item{this_log}{Logical scalar; set to TRUE if the log transformation is used in the decomposition, FALSE otherwise. Default: TRUE}

\item{nfcasts}{Integer scalar; number of projected adjustment factors; should match the \code{nfcasts} argument used to generate the decomposition.}

\item{return_series}{Character scalar; component returned by the routine. Default is "combined"; other possible choices are "seasonal" and "holiday".}

\item{return_xts}{Logical scalar; return projected factors of type \code{return_series} as an \code{xts} time series object. Default is FALSE.
If TRUE, \code{this_x} must be an \code{xts} time series object.}
}
\value{
Array of projected adjustment factors from the fractional airline decomposition
}
\description{
Generates projected adjustment factors from the decomposition of the fractional airline model R routines
}
\examples{
ic_est <-
   rjd3highfreq::fractionalAirlineEstimation(ic_obs, periods=c(365.25/7),
                                             x=ic_default_matrix)
ic_decomp <-
   rjd3highfreq::fractionalAirlineDecomposition(ic_est$model$linearized, 365.25/7, 
                                                stde = TRUE, nfcasts = 104)
ic_xtype <- c(rep('hol', 13), rep('ao', 40))
ic_proj_seasonal <- 
   gen_air_projected_factors(ic_est, ic_decomp,
                      this_xtype = ic_xtype,
                      this_x     = ic_holiday_matrix_fcst,
                      this_log = FALSE, nfcasts = 104)
}
