% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_hybrid_sa_xts.r
\name{gen_hybrid_sa_xts}
\alias{gen_hybrid_sa_xts}
\title{Generate a hybrid seasonal adjustment of an xts time series object}
\usage{
gen_hybrid_sa_xts(
  this_mult_sa,
  this_add_sa,
  this_start_hybrid,
  this_end_hybrid,
  this_week_xts,
  this_year_xts
)
}
\arguments{
\item{this_mult_sa}{xts time series object of a multiplicative seasonal adjustment}

\item{this_add_sa}{xts time series object of an additive seasonal adjustment}

\item{this_start_hybrid}{integer vector of length 2, start of the span where additive adjustments replace multiplicative adjustment, defined as \code{c(year, week)}}

\item{this_end_hybrid}{integer vector of length 2, end of the span where additive adjustments replace multiplicative adjustment, defined as \code{c(year, week)}}

\item{this_week_xts}{Numeric vector; Week of the year for each observation, stored as an \code{xts} time series object.}

\item{this_year_xts}{Numeric vector; Year of each observation, stored as an \code{xts} time series object.}
}
\value{
xts series object with hybrid seasonal adjustment
}
\description{
Generates a "hybrid" seasonal adjustment by replacing a span of a multiplicative seasonal adjustment with an additive adjustment
}
\examples{
cc_hybrid_sa <- gen_hybrid_sa_xts(cc_sa_tc_log, cc_sa_tc_nolog, c(2020,12), c(2020, 52),
                                  cc_week_xts, cc_year_xts)
}
