% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_level_outlier_matrix.r
\name{gen_level_outlier_matrix}
\alias{gen_level_outlier_matrix}
\title{Generate level change regression matrix}
\usage{
gen_level_outlier_matrix(
  outlier_dates,
  this_week,
  this_year,
  forecast = 0,
  air_name = TRUE,
  x13type = TRUE,
  return_xts = FALSE
)
}
\arguments{
\item{outlier_dates}{Integer matrix - matrix of dates for LS outliers}

\item{this_week}{Numeric vector; Week of the year for each observation.}

\item{this_year}{Numeric vector; Year of each observation.}

\item{forecast}{Numeric scalar; Number of forecasts. Default: 0.}

\item{air_name}{Logical scalar; If TRUE, names are formatted as they are in the fractional airline routine;
otherwise format them as in movereg. Default: TRUE}

\item{x13type}{Logical scalar; Indicates if level change outlier is defined as in X-13ARIMA-SEATS. Default: TRUE.}

\item{return_xts}{Logical scalar; return matrix as an \code{xts} time series object. Default is FALSE.
If TRUE, \code{this_week} and \code{this_year} should be \code{xts} time series objects.}
}
\value{
Generate matrix of level change outlier regressors, with column names that describe the individual regressors
}
\description{
Generates a regression matrix of LS (level change) outliers to be included in weekly modeling routines
}
\examples{
ic_level_dates <-
   matrix(c(12, 2020, 13, 2020), ncol=2, byrow=TRUE)
ic_level_matrix <-
   gen_level_outlier_matrix(ic_level_dates, ic_week, ic_year, 0,
                            return_xts = FALSE)
cc_level_dates <-
   matrix(c(12, 2020, 13, 2020), ncol=2, byrow=TRUE)
cc_level_matrix_xts <-
   gen_level_outlier_matrix(cc_level_dates, cc_week_xts, cc_year_xts, 0,
                            return_xts = TRUE)
}
