% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_movereg_holiday_xts.r
\name{gen_movereg_holiday_xts}
\alias{gen_movereg_holiday_xts}
\title{Generate Movreg holiday regression matrix, using xts series as input}
\usage{
gen_movereg_holiday_xts(
  hol_n,
  hol_index,
  hol_wt,
  hol_type,
  this_week,
  this_year,
  return_xts = FALSE
)
}
\arguments{
\item{hol_n}{Numeric scalar; number of weights for this particular holiday}

\item{hol_index}{Numeric scalar; position of holiday within the weight vector}

\item{hol_wt}{Numeric vector; weight vector for holiday effect}

\item{hol_type}{Character scalar; Type of holiday (\code{'newyear'} for New Years Day, \code{'mlk'} for MLK Observance, \code{'president'} for President's Day, \code{'easter'} for Easter, \code{'memorial'} for Memorial Day, \code{'july4'} for Independence Day, \code{'labor'} for Labor Day, \code{'columbus'} for Columbus Day, \code{'veteran'} for Veterans Day, \code{'thanks'} for Thanksgiving).}

\item{this_week}{Numeric vector; Week of the year for each observation, stored as an \code{xts} time series object.}

\item{this_year}{Numeric vector; Year of each observation, stored as an \code{xts} time series object.}

\item{return_xts}{Logical scalar; return matrix as an \code{xts} time series object. Default is FALSE.}
}
\value{
Generate matrix of holiday regressors as defined within the BLS weekly adjustment program Movereg
}
\description{
Generates a matrix with holiday regressors as defined within the BLS weekly adjustment program Movereg
}
\examples{
ic_easter_xts  <-
  gen_movereg_holiday_xts(hol_n = 8,
                          hol_index = 8,
                          hol_wt = c(1, 0, 0, 0, 0, 0, 0, 0),
                          hol_type = 'easter',
                          this_week = ic_week_xts,
                          this_year = ic_year_xts)
ic_labor_xts  <-
  gen_movereg_holiday_xts(hol_n = 2,
                          hol_index = 2,
                          hol_wt = c(0, 1),
                          hol_type = 'labor',
                          this_week = ic_week_xts,
                          this_year = ic_year_xts)
ic_thanksgiving_xts <-
  gen_movereg_holiday_xts(hol_n = 1,
                          hol_index = 1,
                          hol_wt = array(1, dim=1),
                          hol_type = 'thanksgiving',
                          this_week = ic_week_xts,
                          this_year = ic_year_xts)
ic_holiday_matrix_reduced_xts <- 
   cbind(ic_easter_xts, ic_labor_xts, ic_thanksgiving_xts)
colnames(ic_holiday_matrix_reduced_xts) <- c('easter', 'labor', 'thanksgiving')
}
