% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tc_outlier_matrix.r
\name{gen_tc_outlier_matrix}
\alias{gen_tc_outlier_matrix}
\title{Generate temporary change outlier regression matrix}
\usage{
gen_tc_outlier_matrix(
  outlier_dates,
  this_week,
  this_year,
  forecast = 0,
  air_name = TRUE,
  this_freq = NULL,
  tc_alpha = NULL,
  return_xts = FALSE
)
}
\arguments{
\item{outlier_dates}{Integer matrix - matrix of dates for TC outliers}

\item{this_week}{Numeric vector; Week of the year for each observation.}

\item{this_year}{Numeric vector; Year of each observation.}

\item{forecast}{Numeric scalar; Number of forecasts. Default: 0.}

\item{air_name}{Logical scalar; If TRUE, names are formatted as they are in the fractional airline routine;
otherwise format them as in movereg. Default: TRUE}

\item{this_freq}{Numeric scalar; frequency of time series. Default: 365.25/7, for a weekly series}

\item{tc_alpha}{Numeric scalar; Rate of decay for the TC outlier.
Default: will be computed as in \code{X-13ARIMA-SEATS} for a weekly series}

\item{return_xts}{Logical scalar; return matrix as an \code{xts} time series object. Default is FALSE.
If TRUE, \code{this_week} and \code{this_year} should be \code{xts} time series objects.}
}
\value{
Generate matrix of temporary change outlier regressors, with column names that describe the individual regressors
}
\description{
Generates a regression matrix of TC (temporary change) outliers to be included in weekly modeling routines
}
\examples{
ic_tc_dates <-
   matrix(c(13, 2020), ncol=2, byrow=TRUE)
ic_tc_matrix <-
   gen_level_outlier_matrix(ic_tc_dates, ic_week, ic_year, 0,
                            return_xts = FALSE)
cc_tc_dates <-
   matrix(c(13, 2020), ncol=2, byrow=TRUE)
cc_tc_matrix_xts <-
   gen_level_outlier_matrix(cc_tc_dates, cc_week_xts, cc_year_xts, 0,
                            return_xts = TRUE)
}
