% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_day.r
\docType{data}
\name{ic_day}
\alias{ic_day}
\title{Day vector for Initial Unemployment Claims, weekly (ic)}
\format{
A tis time series object for the day of IC from February 1, 2003 to January 30, 2021
}
\usage{
ic_day
}
\description{
A tis time series object the day for every observation of the weekly initial unemployment claims released by the Department of Labor
}
\keyword{datasets}
