% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_end.r
\docType{data}
\name{ic_end}
\alias{ic_end}
\title{Ending date for Initial Unemployment Claims, weekly (ic)}
\format{
A numeric vector of length 2.
\describe{
\item{first element}{starting year}
\item{second element}{starting week}
}
}
\source{
\url{https://www.dol.gov/ui/data.pdf}
}
\usage{
ic_end
}
\description{
A numeric vector of length 2 with the end date of the weekly initial unemployment claims released by the Department of Labor
}
\keyword{datasets}
