% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_month.r
\docType{data}
\name{ic_month}
\alias{ic_month}
\title{Month vector for Initial Unemployment Claims, weekly (ic)}
\format{
A tis time series object for the month of IC from February 1, 2003 to January 30, 2021
}
\usage{
ic_month
}
\description{
A tis time series object the month for every observation of the weekly initial unemployment claims released by the Department of Labor
}
\keyword{datasets}
