% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_obs.r
\docType{data}
\name{ic_obs}
\alias{ic_obs}
\title{Initial Unemployment Claims, weekly (ic)}
\format{
A tis time series object for IC from February 1, 2003 to January 30, 2021
}
\source{
\url{https://www.dol.gov/ui/data.pdf}
}
\usage{
ic_obs
}
\description{
A time series object with weekly initial unemployment claims released by the Department of Labor
}
\keyword{datasets}
