% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_outlier.r
\docType{data}
\name{ic_outlier}
\alias{ic_outlier}
\title{Seasonal Factors for Initial Unemployment Claims, weekly (ic)}
\format{
A tis time series object from February 1, 2003 to January 30, 2021
}
\usage{
ic_outlier
}
\description{
A time series object with seasonal factors generated for weekly initial unemployment
claims released by the Department of Labor
}
\keyword{datasets}
