% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_sf_auto.r
\docType{data}
\name{ic_sf_auto}
\alias{ic_sf_auto}
\title{Seasonal Factors for Initial Unemployment Claims, weekly (ic, automatic outlier identification)}
\format{
A tis time series object from February 1, 2003 to January 30, 2021
}
\usage{
ic_sf_auto
}
\description{
A time series object with seasonal factors generated for weekly initial unemployment
claims from a model that includes automatically identified outliers
}
\keyword{datasets}
