% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_sf_auto_ljung.r
\docType{data}
\name{ic_sf_auto_ljung}
\alias{ic_sf_auto_ljung}
\title{Seasonal Factors for Initial Unemployment Claims, weekly (ic, automatic outlier identification, ljung)}
\format{
A tis time series object from February 1, 2003 to January 30, 2021
}
\usage{
ic_sf_auto_ljung
}
\description{
A time series object with seasonal factors generated for weekly initial unemployment
claims from a model that includes automatically identified outliers with a critical value
determined by an algorithm developed by Greta Ljung
}
\keyword{datasets}
