% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_sf_tc.r
\docType{data}
\name{ic_sf_tc}
\alias{ic_sf_tc}
\title{Seasonal Factors for Initial Unemployment Claims, weekly (automatic outlier identification, TC outlier specified)}
\format{
A tis time series object from February 1, 2003 to January 30, 2021
}
\usage{
ic_sf_tc
}
\description{
A time series object with seasonal factors generated for weekly initial unemployment
claims from a model that includes automatically identified outliers with a temporary
change outlier (TC) specified for the 13th week of 2020
}
\keyword{datasets}
