% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_week.r
\docType{data}
\name{ic_week}
\alias{ic_week}
\title{Week vector for Initial Unemployment Claims, weekly (ic)}
\format{
A tis time series object for IC from February 1, 2003 to January 30, 2021
}
\source{
\url{https://www.dol.gov/ui/data.pdf}
}
\usage{
ic_week
}
\description{
A tis time series object the week number for every observation of the weekly initial unemployment claims released by the Department of Labor
}
\keyword{datasets}
