% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_xts.r
\docType{data}
\name{ic_xts}
\alias{ic_xts}
\title{Initial Unemployment Claims, weekly (ic) in xts format}
\format{
An xts time series object for IC from February 1, 2003 to January 30, 2021
}
\source{
\url{https://www.dol.gov/ui/data.pdf}
}
\usage{
ic_xts
}
\description{
An xts time series object with weekly initial unemployment claims released by the Department of Labor
}
\keyword{datasets}
