% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_xts_dates.r
\docType{data}
\name{ic_xts_dates}
\alias{ic_xts_dates}
\title{Date Index used in the xts object for Initial Unemployment Claims, weekly (ic)}
\format{
An vector of weekly dates from February 1, 2003 to January 30, 2021
}
\source{
\url{https://www.dol.gov/ui/data.pdf}
}
\usage{
ic_xts_dates
}
\description{
A vector of dates used as an index for the xts object of weekly initial unemployment claims released by the Department of Labor
}
\keyword{datasets}
