% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_year.r
\docType{data}
\name{ic_year}
\alias{ic_year}
\title{Year vector for Initial Unemployment Claims, weekly (ic)}
\format{
A tis time series object for the year of IC from February 1, 2003 to January 30, 2021
}
\usage{
ic_year
}
\description{
A tis time series object the year for every observation of the weekly initial unemployment claims released by the Department of Labor
}
\keyword{datasets}
