% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobian_trans_adj.r
\name{jacobian_trans_adj}
\alias{jacobian_trans_adj}
\title{Jacobian transformation adjustment}
\usage{
jacobian_trans_adj(
  Y,
  Nspobs,
  Nintvl = 2,
  Adj = NULL,
  Logit = FALSE,
  Trans = TRUE,
  Lam = 0
)
}
\arguments{
\item{Y}{numeric vector, original time series.}

\item{Nspobs}{Integer scalar, Length of \code{Y}.}

\item{Nintvl}{Integer scalar, Number of differences in model used. Default is 2.}

\item{Adj}{numeric vector, prior adjustment factor time series. Default is NULL, which indicates there is no prior adjustment.}

\item{Logit}{logical scalar, if TRUE the logit transformation is used. Default is FALSE.}

\item{Trans}{logical scalar, if TRUE a Box-Cox transform is used. Default is TRUE}

\item{Lam}{numeric scalar, Box-Cox transformation parameter. Default is 0 (Log transform).}
}
\value{
Jacobean adjustment to the likelihood of estimated models to allow AIC and other likelihood statistics to be compared for different transformations.
}
\description{
compute the Jacobian adjustment for transformations as done in the X-13ARIMA-SEATS program
}
\examples{
Nspobs <- length(ic_obs)
ic_jacadj_log   <- jacobian_trans_adj(ic_obs, Nspobs, 2, Trans = TRUE, Lam = 0.0)   
}
