% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_date.r
\name{match_date}
\alias{match_date}
\title{Match a specific date}
\usage{
match_date(x, date_string)
}
\arguments{
\item{x}{Numeric tis object.}

\item{date_string}{Character string; Year, month and day of the date (Example - January 1, 1990 is '19900101').}
}
\value{
An indicator variable where the week contains the date entered = 1, 0 otherwise.
}
\description{
Create an indicator variables for a specific date (year, month, day) for a weekly time series as defined in \code{tis}.
}
\examples{
pandemic_start <- match_date(ic_week, '20200317')
}
