% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_month_day.r
\name{match_month_day}
\alias{match_month_day}
\title{Match a specific month and day}
\usage{
match_month_day(x, date_string)
}
\arguments{
\item{x}{Numeric tis object.}

\item{date_string}{Character string; Month and day of the date (Example - January 1 is '0101').}
}
\value{
An indicator variable where the week that contains the date entered = 1, 0 otherwise.
}
\description{
Create an indicator variables for a specific month and day.
}
\examples{
july4_wed <- match_month_day(ic_week, '0707')
}
