% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_month_day_xts.r
\name{match_month_day_xts}
\alias{match_month_day_xts}
\title{Match a specific month and day}
\usage{
match_month_day_xts(x, date_string, return_xts = FALSE)
}
\arguments{
\item{x}{Numeric xts object.}

\item{date_string}{Character string; Month and day of the date (Example - January 1 is '0101').}

\item{return_xts}{Logical scalar; return matrix as an \code{xts} time series object. Default is FALSE.}
}
\value{
An indicator variable where the week that contains the date entered = 1, 0 otherwise.
}
\description{
Create an indicator variables for a specific month and day.
}
\examples{
july4_wed_xts <- match_month_day_xts(ic_week_xts, '0707')
}
