% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_week_xts.r
\name{match_week_xts}
\alias{match_week_xts}
\title{Match a specific week}
\usage{
match_week_xts(x_week, week_number, return_xts = FALSE)
}
\arguments{
\item{x_week}{Numeric xts object with week of each observation.}

\item{week_number}{Numeric scalar; Week of the year to match.}

\item{return_xts}{Logical scalar; return matrix as an \code{xts} time series object. Default is FALSE.}
}
\value{
An indicator variable where the observation that matches the week entered = 1, 0 otherwise.
}
\description{
Create an indicator variable for a specific week.
}
\examples{
xmas_w53_xts <- match_week_xts(ic_week_xts, 53)
}
