% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxplot_weekly.r
\name{plot_boxplot_weekly}
\alias{plot_boxplot_weekly}
\title{Generate weekly boxplots}
\usage{
plot_boxplot_weekly(
  this_factor,
  this_week,
  forecast = 0,
  this_range = NULL,
  this_type = "Irregular",
  this_title
)
}
\arguments{
\item{this_factor}{Numeric vector of factors from a weekly seasonal adjustment.}

\item{this_week}{Numeric vector; Week of the year to match.}

\item{forecast}{Numeric scalar; Number of forecasts. Default: 0.}

\item{this_range}{Numeric scalar; Range of the seasonal factors to be plotted. Default: range taken from the observations being plotted.}

\item{this_type}{Character string; type of factor being plotted. Default: 'Irregular'}

\item{this_title}{Character string; Main title for plot.}
}
\value{
Generate a mulit-panel plot of boxplots of factors by week. Need to have \code{plot_boxplot_single} loaded to run.
}
\description{
Generate a multi-panel plot of boxplots by week
}
\examples{
ic_sf_range <- 
   range(ic_sf, ic_sf_auto)
plot_boxplot_weekly(ic_sf, ic_week, 
                    forecast = 0,
                    this_range = ic_sf_range,
                    this_type  = "Seasonal",
                    this_title = 'Initial Claims (default)')
plot_boxplot_weekly(ic_sf_auto, ic_week, 
                    forecast = 0,
                    this_range = ic_sf_range, 
                    this_type  = "Seasonal",
                    this_title = 'Initial Claims (auto outliers)')
}
