% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sf_weekly.r
\name{plot_sf_weekly}
\alias{plot_sf_weekly}
\title{Plot factors by week}
\usage{
plot_sf_weekly(
  this_sf,
  this_week,
  forecast = 0,
  this_trans = TRUE,
  this_range = NULL,
  this_type = "Seasonal",
  this_title
)
}
\arguments{
\item{this_sf}{Numeric vector of seasonal factors.}

\item{this_week}{Numeric vector; Week of the year to match.}

\item{forecast}{Numeric scalar; Number of forecasts. Default: 0.}

\item{this_trans}{Logical scalar; set to TRUE if the log transformation is used in the decomposition, FALSE otherwise. Default: TRUE}

\item{this_range}{Numeric scalar; Range of the seasonal factors to be plotted. Default: range taken from the observations being plotted.}

\item{this_type}{Character string; type of factor being plotted. Default: 'Seasonal'}

\item{this_title}{Character string; Main title for plot.}
}
\value{
Generate a mulit-panel plot of seasonal factors by week. Need to have \code{plot_sf_weekly_single} loaded to run.
}
\description{
Generate a multi-panel plot of seasonal factors by week
}
\examples{
ic_sf_range <- range(ic_sf, ic_sf_auto, ic_sf_auto_ljung, ic_sf_tc)
plot_sf_weekly(ic_sf, 
               this_week = ic_week,  
               forecast = 0,
               this_range = ic_sf_range,  
               this_trans = FALSE,  
               this_type = 'Combined',
               this_title = 'Initial Claims (default)')
plot_sf_weekly(ic_sf_auto,  
               this_week = ic_week,  
               forecast=0,
               this_range = ic_sf_range,  
               this_trans = FALSE,  
               this_type = 'Combined',
               this_title = 'Initial Claims (auto)')
plot_sf_weekly(ic_sf_auto_ljung,  
               this_week = ic_week,  
               forecast=0,
               this_range = ic_sf_range,  
               this_trans = FALSE,  
               this_type = 'Combined',
               this_title = 'Initial Claims (auto, ljung cv)')
plot_sf_weekly(ic_sf_tc,  
               this_week = ic_week,  
               forecast=0,
               this_range = ic_sf_range,  
               this_trans = FALSE,  
               this_type = 'Combined',
               this_title = 'Initial Claims (fractional airline)')
}
