% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_values_air.r
\name{t_values_air}
\alias{t_values_air}
\title{Generate t-statistics}
\usage{
t_values_air(this_beta, this_cov)
}
\arguments{
\item{this_beta}{Numeric vector; estimated model parameters from the fractional airline model.}

\item{this_cov}{Numeric matrix; estimated covariance matrx from the fractional airline model.}
}
\value{
Numeric vector of t-values
}
\description{
Generate t-statistics from estimates generated from the fractional airline model
}
\examples{
ic_air_est <-
  rjd3highfreq::fractionalAirlineEstimation(ic_obs, periods=c(365.25/7),
                                     x=ic_holiday_matrix,
                                     outliers=c('ao', 'ls'),
                                     criticalValue = 6)
ic_air_t_reg   <- t_values_air(ic_air_est$model$b, ic_air_est$model$bcov)
ic_air_t_arima <- t_values_air(ic_air_est$estimation$parameters,
                               ic_air_est$estimation$covariance)
}
