% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within_month_regressors.r
\name{within_month_regressors}
\alias{within_month_regressors}
\title{Generate within-month effect regression matrix}
\usage{
within_month_regressors(
  this_year,
  this_month,
  this_day,
  l = 30,
  return_xts = TRUE
)
}
\arguments{
\item{this_year}{Numeric vector; Year of each observation.}

\item{this_month}{Numeric vector; Month for each observation.}

\item{this_day}{Numeric vector; Day of the month for each observation.}

\item{l}{Numeric scalar; Number of sin-cos terms to generate, with the total number of regressors being \code{2*l}. Default: 30.}

\item{return_xts}{Logical scalar; return matrix as an \code{xts} time series object. Default is FALSE.}
}
\value{
Generate matrix of within month effects regressors, with column names that describe the individual regressors
}
\description{
Generates a regression matrix of within month effects from Cleveland and Scott to be included in weekly modeling routines
}
\examples{
ic_wm_matrix_1992 <-
   within_month_regressors(ic_year, ic_month, ic_day, return_xts = FALSE)
cc_wm_matrix_xts <-
   within_month_regressors(cc_year_xts, cc_month_xts, cc_day_xts, 
                           return_xts = TRUE)
}
